requirements.txt
pillow
discord.py>=2.3.0
requests>=2.28.0

